Implemented new ArrayDisplayBlock for Stack and Queue question.
Refactored Stack and Queue generator into multiple smaller functions
