+++ settings/questionsSelection.ts
@@ -14,8 +14,8 @@ import { NormalForms } from "@shared/question-generators/propositional-logic/nor
 import { Satisfiability } from "@shared/question-generators/propositional-logic/satisfiability.ts"
 import { RecursionFormula } from "@shared/question-generators/recursion/formula"
 import { RecurrenceMaster } from "@shared/question-generators/recursion/recurrenceMaster"
-import { queueQuestion } from "@shared/question-generators/StackQueue/QueueGenerator.ts"
-import { stackQuestion } from "@shared/question-generators/StackQueue/StackGenerator.ts"
+import { queueQuestion } from "@shared/question-generators/StackQueue/Queue/QueueGenerator.ts"
+import { stackQuestion } from "@shared/question-generators/StackQueue/Stack/StackGenerator.ts"
 import { Loops } from "@shared/question-generators/time/loops"
 
 export const DEFAULT_IMAGE = new URL("../../assets/images/skill-default.jpg", import.meta.url)
