+++ shared/src/question-generators/StackQueue/Stack/Stack.ts
